+++ examples/storybook/stories/Plugins/FileUpload/FileUploadEditor.tsx
@@ -1,6 +1,5 @@
 import React, { FunctionComponent } from 'react';
 import { DraftContent, RicosEditor } from 'ricos-editor';
-import { RichContentEditor } from 'wix-rich-content-editor';
 import { pluginFileUpload } from 'wix-rich-content-plugin-file-upload';
 
 const FileUploadEditor: FunctionComponent<{
@@ -10,11 +9,7 @@ const FileUploadEditor: FunctionComponent<{
   const config = {
     onFileSelected: onFilesChange,
   };
-  return (
-    <RicosEditor plugins={[pluginFileUpload(config)]} content={content}>
-      <RichContentEditor />
-    </RicosEditor>
-  );
+  return <RicosEditor plugins={[pluginFileUpload(config)]} content={content} />;
 };
 
 export default FileUploadEditor;