+++ packages/plugin-divider/web/src/tiptap/tiptap.ts
@@ -1,6 +1,7 @@
 import dividerDataDefaults from 'ricos-schema/dist/statics/divider.defaults.json';
 import { CreateRicosExtensions } from 'wix-tiptap-editor';
 import { Divider as Component } from './component';
+import { TIPTAP_DIVIDER_TYPE } from 'ricos-content';
 
 export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
   {
@@ -8,7 +9,7 @@ export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
     Component,
     componentDataDefaults: { ...dividerDataDefaults, id: '' },
     createExtensionConfig: () => ({
-      name: 'divider',
+      name: TIPTAP_DIVIDER_TYPE,
       defaultOptions,
     }),
   },