+++ packages/plugin-giphy/web/src/tiptap/tiptap.ts
@@ -1,6 +1,7 @@
-import { CreateRicosExtensions, TIPTAP_GIF_TYPE } from 'wix-tiptap-editor';
+import { CreateRicosExtensions } from 'wix-tiptap-editor';
 import { Gif as Component } from './component';
 import gifDataDefaults from 'ricos-schema/dist/statics/gif.defaults.json';
+import { TIPTAP_GIF_TYPE } from 'ricos-content';
 
 const name = TIPTAP_GIF_TYPE;