+++ packages/plugin-image/web/src/tiptap/tiptap.ts
@@ -1,6 +1,7 @@
 import imageDataDefaults from 'ricos-schema/dist/statics/image.defaults.json';
 import { CreateRicosExtensions, TIPTAP_IMAGE_TYPE } from 'wix-tiptap-editor';
 import { Image as Component } from './component';
+import { TIPTAP_IMAGE_TYPE } from 'ricos-content';
 
 declare module '@tiptap/core' {
   interface Commands<ReturnType> {